// part9/patientor/src/types.ts
export interface Diagnosis {
	code: string;
	name: string;
	latin?: string;
}

export enum Gender {
	Male = "male",
	Female = "female",
	Other = "other"
}

export interface BaseEntry {
	id: string;
	description: string;
	date: string;
	specialist: string;
	diagnosisCodes?: Array<Diagnosis['code']>;
}

export enum HealthCheckRating {
	'Healthy' = 0,
	'LowRisk' = 1,
	'HighRisk' = 2,
	'CriticalRisk' = 3
}

export interface HealthCheckEntry extends BaseEntry {
	type: 'HealthCheck';
	healthCheckRating: HealthCheckRating;
}

export interface OccupationalHealthCareEntry extends BaseEntry {
	type: 'OccupationalHealthcare';
	employerName: string;
	sickLeave?: { startDate: string; endDate: string; };
}

export interface HospitalEntry extends BaseEntry {
	type: 'Hospital';
	discharge: { date: string; criteria: string; };
}

export type Entry =
	| HospitalEntry
	| OccupationalHealthCareEntry
	| HealthCheckEntry;

export const EntryOptions = ['HealthCheck', 'OccupationalHealthcare', 'Hospital'];

export interface Patient {
	id: string;
	name: string;
	occupation: string;
	gender: Gender;
	ssn?: string;
	dateOfBirth?: string;
	entries?: Entry[];
}

export interface PatientDetailed extends Patient {
	ssn: string;
	dateOfBirth: string;
	entries: Entry[];
}

export const assertNever = (value: never): never => {
	throw new Error(
		`Unhandled discriminated union member: ${JSON.stringify(value)}`
	);
};